+++ packages/__tests__/src/1-kernel/modules/command-multi.ts
@@ -1,4 +1,4 @@
-import { bindingCommand } from '@aurelia/runtime-html';
+import { bindingCommand } from '@aurelia/template-compiler';
 
 @bindingCommand('cmd')
 export class Cmd {}
