+++ packages/compat-v1/src/compat-event.ts
@@ -2,12 +2,8 @@ import { DI, IContainer, resolve } from '@aurelia/kernel';
 import { IObserverLocatorBasedConnectable } from '@aurelia/runtime';
 import {
   AppTask,
-  type BindingCommandInstance,
-  type ICommandBuildInfo,
   IEventTarget,
   type IHydratableController,
-  IInstruction,
-  InstructionType,
   IRenderer,
   mixinAstEvaluator,
   mixinUseScope,
@@ -15,7 +11,6 @@ import {
   renderer,
   IPlatform,
   IListenerBindingOptions,
-  type BindingCommandStaticAuDefinition,
   astBind,
   astEvaluate,
   astUnbind,
@@ -23,6 +18,12 @@ import {
   type IBinding,
   type Scope,
 } from '@aurelia/runtime-html';
+import {
+  type BindingCommandInstance,
+  type ICommandBuildInfo,
+  type IInstruction,
+  type BindingCommandStaticAuDefinition,
+} from '@aurelia/template-compiler';
 import { createLookup, ensureExpression, etIsFunction, isFunction } from './utilities';
 import { IExpressionParser, IsBindingBehavior } from '@aurelia/expression-parser';
 
@@ -99,7 +100,7 @@ export const ListenerBindingRenderer = /*@__PURE__*/ renderer(class ListenerBind
 }, null!);
 
 export class DelegateBindingInstruction {
-  public readonly type = InstructionType.listenerBinding;
+  public readonly type = 'dl';
 
   public constructor(
     public from: string | IsBindingBehavior,
