+++ packages/state/src/state-templating.ts
@@ -4,17 +4,19 @@ import {
   IObserverLocator,
 } from '@aurelia/runtime';
 import {
-  AttrSyntax,
-  IAttrMapper,
   IHydratableController,
   IPlatform,
   renderer,
+  type IRenderer,
+} from '@aurelia/runtime-html';
+import {
+  AttrSyntax,
+  IAttrMapper,
   type BindingCommandInstance,
   type ICommandBuildInfo,
   type IInstruction,
-  type IRenderer,
-  BindingCommandStaticAuDefinition
-} from '@aurelia/runtime-html';
+  type BindingCommandStaticAuDefinition
+} from '@aurelia/template-compiler';
 import { IStore } from './interfaces';
 import { StateBinding } from './state-binding';
 import { StateDispatchBinding } from './state-dispatch-binding';
