+++ packages/ui-virtualization/package.json
@@ -59,6 +59,7 @@
     "@aurelia/platform-browser": "2.0.0-beta.15",
     "@aurelia/expression-parser": "2.0.0-beta.15",
     "@aurelia/runtime": "2.0.0-beta.15",
+    "@aurelia/template-compiler": "2.0.0-beta.15",
     "@aurelia/runtime-html": "2.0.0-beta.15"
   },
   "devDependencies": {
